compat/timegm: new compat function for systems lacking timegm()
The implementation used here is taken from Samba, since discussions on
their mailing list indicate that the "portable version" in timegm(3)
does not actually work on Solaris.

This commit also adds the infrastructure required for compat functions
in CGit itself.  Mostly we get everything we need from Git, but timegm
is not used in Git so there is no reason for them to carry a
compatibility version.

Signed-off-by: John Keeping <john@keeping.me.uk>