Check SHA256 sum of git-$VER.tar.gz after downloading
This requires that we save the downloaded file explicitly rather than
piping it straight to tar, but that is advisable anyway since it allows
us to check the exit status of curl and make sure that we have
downloaded the file successfully.

Since not all systems have `sha256sum` available, we fall back to
displaying a warning when it is not found.  This is still an improvement
since it lets people know that they may wish to take extra steps to
verify the version of Git that has been downloaded.

Also add a test to make sure we don't forget to update the file when
updating our Git version in the future.

Signed-off-by: John Keeping <john@keeping.me.uk>