ui-blame: add blame UI
This is disabled by default and needs to be turned on in the config
file.

Because libgit.a does not include the blame implementation (which lives
in git/builtin/blame.c), this is implemented by executing a git-blame
subprocess and parsing its output.  Given how expensive the blame
operation itself is, the overhead of using a separate process should not
be noticeable.