filter: introduce cgit_open_email_filter() wrapper
We provide email addresses to the email filter surrounded by angle
brackets, but we will soon remove these in our internal representation.
Introduce a wrapper so that we only have to add them in one place.

Signed-off-by: John Keeping <john@keeping.me.uk>
---
It might be nice for this to be symmetric with a new
cgit_close_email_filter() thin wrapper as well, but I haven't bothered
with that in this first pass.