global: avoid hardcoded SHA1 sizes
Use the GIT_MAX_HEXSZ value so that we will pick up any changes in
libgit.a to support larger object IDs using has algorithms other than
SHA1.

I'm sure there's more to do for CGit to support other hash algorithms,
but this catches everywhere we use "20" or "40" to refer to a hash
length as either raw or hex encoded.

Signed-off-by: John Keeping <john@keeping.me.uk>