cache: avoid truncating ssize_t to int
This can't happen because xread() will return a value between -1 and
CACHE_BUFSIZE = 4096 which will fit in the range of int, but let's avoid
a potential warning by using the correct type.

While here, clarify the use of bufkeylen by reducing its scope so that
we no longer need a placeholder value.  This means we can change it to
an unsigned type and hopefully shut up a Coverity warning about
undefined overflow from the "bufkeylen + 1".

Signed-off-by: John Keeping <john@keeping.me.uk>