tests: document handling of repository names with spaces
t0103-log.sh has a test to check that we don't properly escape spaces in
repository URLs, but no other tests do.

It turns out that the handling of this case in the index page is broken,
so mark this as test_expect_failure while adding successful tests for
all other pages.

Signed-off-by: John Keeping <john@keeping.me.uk>