ui-{repolist,atom}: fix repository URL escaping
Remove cgit_repourl() and replace it with a new function that prints a
suitable escaped version of the repository URL designed for use in an
attribute value.  There are only two places that use this at the moment
and both want the value to be used in attribute context.

We could introduce new escaping functions to let us keep cgit_repourl()
and return a suitable escaped string, but there's no need for us to be
doing memory management here so this ends up simplifying the call sites.

Signed-off-by: John Keeping <john@keeping.me.uk>