[RFC] Support compression for patch and rawdiff
This needs splitting up into a series of a few patches (I think there's
about four hiding in here), but I'm sending it now to see if this is
roughly what's needed.

The idea is to support extensions on page names (for pages that opt in)
and it's implemented for patch and rawdiff at the moment with support
for GZIP compressing the response body.

For example, given the URL:

	https://git.zx2c4.com/cgit/patch/?id=8e9ddd21a50beb9fd660cf6cd6a583234924b932

with this change you can add ".gz" to the end of "patch":

	https://git.zx2c4.com/cgit/patch.gz/?id=8e9ddd21a50beb9fd660cf6cd6a583234924b932

and the content will be returned compressed with gzip.

I've only added gzip for now, but the setup is table driven so adding
more compression filters should be straightforward.

Signed-off-by: John Keeping <john@keeping.me.uk>