cgitrc.5: fix continuation paragraphs
Asciidoc has a slightly strange requirement for continuing an indented
block with a new paragraph, which requires a "+" line linking the two
paragraphs and that the follow-on paragraph start in column zero.  If
the continuation paragraph is simply indented after a blank line, it is
interpreted as code and formatted in a fixed-width font.

Fix a couple of places in cgitrc.5 which get thisi wrong.

Signed-off-by: John Keeping <john@keeping.me.uk>