cgit_query: rename sha1 and sha2 fields
These come from the "id" and "id2" parameters and nowhere do we require
the content to be a SHA-1 (or even a hash in most places), so let's
rename the internal variables to match the query parameters.

This helps to easily identify the remaining places where we do have a
dependency on SHA-1 via a simple grep.

Signed-off-by: John Keeping <john@keeping.me.uk>